Session service file creation feature branch

+++ backend_po/session_service/src/main/java/org/session_service/services/SessionService.java
@@ -1,23 +1,32 @@
 package org.session_service.services;
 
+import jakarta.servlet.http.HttpSession;
 import org.session_service.DTO.CreateSessionRequest;
 import org.session_service.DTO.DeleteSessionRequest;
 import org.session_service.entities.SessionEntity;
 import org.session_service.repositories.SessionRepository;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.text.ParseException;
+import java.util.HashMap;
+import java.util.Map;
 
 @Service
 public class SessionService {
 
     private final SessionRepository sessionRepository;
-    private SessionCreationService sessionCreationService;
+    private final SessionCreationService sessionCreationService;
+    private final RequestSendingService requestSendingService;
+    //TODO create a batter way to configure addresses
+    private final String fileServiceAddress = "http://localhost:8080";
 
-    public SessionService(SessionRepository sessionRepository, SessionCreationService sessionCreationService) {
+    public SessionService(SessionRepository sessionRepository, SessionCreationService sessionCreationService, RequestSendingService requestSendingService, HttpSession httpSession) {
         this.sessionRepository = sessionRepository;
         this.sessionCreationService = sessionCreationService;
+        this.requestSendingService = requestSendingService;
     }
 
     public Integer JoinSession(String sessionId) throws IllegalArgumentException {
@@ -30,10 +39,32 @@ public class SessionService {
         return sessionEntity.getSessionId();
     }
 
+    public String getSessionFileId(String sessionId) throws IllegalArgumentException {
+        //session id format is already validated in controller
+        SessionEntity sessionEntity = sessionRepository.findSessionById(Integer.parseInt(sessionId));;
+        if(sessionEntity == null) {
+            throw new IllegalArgumentException("session with this id doesn't exist");
+        }
+
+        return sessionEntity.getSessionFileId();
+    }
+
     public Integer saveSession(CreateSessionRequest createSessionRequest) throws IllegalArgumentException {
         SessionEntity sessionEntity;
         try {
-            sessionEntity = sessionCreationService.createSessionEntity(createSessionRequest.getUserId(), createSessionRequest.getSessionFileId());
+            ResponseEntity<Map> fileServiceResponse = requestSendingService.sendPostRequest(fileServiceAddress+"/fileAPI/createNewFile", null);
+
+            if(fileServiceResponse.getStatusCode() == HttpStatus.BAD_REQUEST || fileServiceResponse.getStatusCode() != HttpStatus.OK){
+                throw new IllegalArgumentException((fileServiceResponse.getBody().get("message") == null ? "invalid response from file service" : "file service answered with error: "+fileServiceResponse.getBody().get("message").toString()));
+            }
+
+            if(fileServiceResponse.getBody() == null || fileServiceResponse.getBody().get("fileId") == null) {
+                throw new IllegalArgumentException("file service answered with invalid request");
+            }
+
+            String sessionFileId = (String) fileServiceResponse.getBody().get("fileId");
+
+            sessionEntity = sessionCreationService.createSessionEntity(createSessionRequest.getUserId(), sessionFileId);
         }catch (RuntimeException e){
             throw new IllegalArgumentException(e.getMessage());
         }
@@ -51,6 +82,16 @@ public class SessionService {
         if(sessionEntity == null) {
             throw new IllegalArgumentException("session with this id doesn't exist");
         }
+
+        String sessionFileId = sessionEntity.getSessionFileId();
+        Map<String, String> fileDeleteRequest = new HashMap<>();
+        fileDeleteRequest.put("fileId", sessionFileId);
+
+        ResponseEntity<Map> response = requestSendingService.sendPostRequestAsMap(fileServiceAddress+"/fileAPI/deleteFile", fileDeleteRequest);
+        if(response.getStatusCode() == HttpStatus.BAD_REQUEST || response.getStatusCode() != HttpStatus.OK) {
+            throw new IllegalArgumentException((response.getBody().get("message") == null ? "invalid response from file service" : "file service answered with error: "+response.getBody().get("message").toString()));
+        }
+
         sessionRepository.delete(sessionEntity);
     }
 }
